[#25] 이미지 업로드 버튼 기능 컴포넌트로 세분화
